+++ exonum-java-binding/CHANGELOG.md
@@ -26,6 +26,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `CheckedProof#getIndexHash`, `ProofListIndexProxy#getIndexHash` and
   `ProofMapIndexProxy#getIndexHash` accordingly.
 
+### Removed
+- Internal list proof representation abstraction `ListProof` is replaced with
+  `UncheckedListProofAdapter`. (#1089)
+
 ## [0.7.0] - 2019-07-17
 
 ### Overview